+++ 01 Operator_Overloading/01 OperatorOverloading/OperatorOverloading.Host2/App.config
@@ -3,4 +3,7 @@
     <startup> 
         <supportedRuntime version="v4.0" sku=".NETFramework,Version=v4.5" />
     </startup>
+  <appSettings>
+    <add key="Path" value="C:/Users/arsalang/Desktop/json.txt"/>
+  </appSettings>
 </configuration>
